+++ chatwoot_client/models/account_summary.py
@@ -30,10 +30,10 @@ class AccountSummary(object):
     swagger_types = {
         'avg_first_response_time': 'str',
         'avg_resolution_time': 'str',
-        'conversations_count': 'float',
-        'incoming_messages_count': 'float',
-        'outgoing_messages_count': 'float',
-        'resolutions_count': 'float',
+        'conversations_count': 'int',
+        'incoming_messages_count': 'int',
+        'outgoing_messages_count': 'int',
+        'resolutions_count': 'int',
         'previous': 'AccountSummaryPrevious'
     }
 
@@ -120,7 +120,7 @@ class AccountSummary(object):
 
 
         :return: The conversations_count of this AccountSummary.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._conversations_count
 
@@ -130,7 +130,7 @@ class AccountSummary(object):
 
 
         :param conversations_count: The conversations_count of this AccountSummary.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._conversations_count = conversations_count
@@ -141,7 +141,7 @@ class AccountSummary(object):
 
 
         :return: The incoming_messages_count of this AccountSummary.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._incoming_messages_count
 
@@ -151,7 +151,7 @@ class AccountSummary(object):
 
 
         :param incoming_messages_count: The incoming_messages_count of this AccountSummary.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._incoming_messages_count = incoming_messages_count
@@ -162,7 +162,7 @@ class AccountSummary(object):
 
 
         :return: The outgoing_messages_count of this AccountSummary.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._outgoing_messages_count
 
@@ -172,7 +172,7 @@ class AccountSummary(object):
 
 
         :param outgoing_messages_count: The outgoing_messages_count of this AccountSummary.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._outgoing_messages_count = outgoing_messages_count
@@ -183,7 +183,7 @@ class AccountSummary(object):
 
 
         :return: The resolutions_count of this AccountSummary.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._resolutions_count
 
@@ -193,7 +193,7 @@ class AccountSummary(object):
 
 
         :param resolutions_count: The resolutions_count of this AccountSummary.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._resolutions_count = resolutions_count
