+++ chatwoot_client/models/conversation_id_assignments_body.py
@@ -28,8 +28,8 @@ class ConversationIdAssignmentsBody(object):
                             and the value is json key in definition.
     """
     swagger_types = {
-        'assignee_id': 'float',
-        'team_id': 'float'
+        'assignee_id': 'int',
+        'team_id': 'int'
     }
 
     attribute_map = {
@@ -54,7 +54,7 @@ class ConversationIdAssignmentsBody(object):
         Id of the assignee user  # noqa: E501
 
         :return: The assignee_id of this ConversationIdAssignmentsBody.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._assignee_id
 
@@ -65,7 +65,7 @@ class ConversationIdAssignmentsBody(object):
         Id of the assignee user  # noqa: E501
 
         :param assignee_id: The assignee_id of this ConversationIdAssignmentsBody.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._assignee_id = assignee_id
@@ -77,7 +77,7 @@ class ConversationIdAssignmentsBody(object):
         Id of the team. If the assignee_id is present, this param would be ignored  # noqa: E501
 
         :return: The team_id of this ConversationIdAssignmentsBody.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._team_id
 
@@ -88,7 +88,7 @@ class ConversationIdAssignmentsBody(object):
         Id of the team. If the assignee_id is present, this param would be ignored  # noqa: E501
 
         :param team_id: The team_id of this ConversationIdAssignmentsBody.  # noqa: E501
-        :type: float
+        :type: int
         """
 
         self._team_id = team_id
